+++ src/ModixTranslator/Models/TranslationService/DetectedLanguage.cs
@@ -1,0 +1,13 @@
+﻿using System.Text.Json.Serialization;
+
+namespace ModixTranslator.Models.TranslationService
+{
+    public class DetectedLanguage
+    {
+        [JsonPropertyName("langauge")]
+        public string Language { get; set; }
+
+        [JsonPropertyName("score")]
+        public float Score { get; set; }
+    }
+}
