+++ src/ModixTranslator/Models/TranslationService/TranslationResponse.cs
@@ -9,15 +9,6 @@ namespace ModixTranslator.Models.TranslationService
         public List<TranslatedText>? Translations { get; set; }
 
         [JsonPropertyName("detectedLanguage")]
-        public LanguageData? DetectedLanguage { get; set; }
-    }
-
-    public class LanguageData
-    {
-        [JsonPropertyName("langauge")]
-        public string Language { get; set; }
-
-        [JsonPropertyName("score")]
-        public float Score { get; set; }
+        public DetectedLanguage? DetectedLanguage { get; set; }
     }
 }
