+++ src/ModixTranslator/Models/Translator/Translation.cs
@@ -31,17 +31,4 @@ namespace ModixTranslator.Models.Translator
 
         public IEnumerable<string> CodeBlocks { get; set; }
     }
-
-    public class LocalText
-    {
-        public LocalText(string language, string text)
-        {
-            Language = language;
-            Text = text;
-        }
-
-        public string Language { get; set; }
-
-        public string Text { get; set; }
-    }
 }
